backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
dlib
Pillow
numpy
